refactor: ルータの整理(#99)
