+++ pkg/server/router/routeing.go
@@ -1,0 +1,15 @@
+package router
+
+import "github.com/labstack/echo/v4"
+
+func rootRouter(e *echo.Echo) {
+	e.POST("/api/v1/login", authHandler.LoginHandler)
+	api := e.Group("/api/v1")
+	{
+		api.Use(authHandler.TokenMiddlewareHandlerFunc)
+		api.POST("/posts", postHandler.Post)
+		api.GET("/posts/:id", postHandler.FindByID)
+
+		api.GET("/users/:id", userHandler.FindByID)
+	}
+}
