+++ pkg/server/router/start.go
@@ -9,22 +9,16 @@ import (
 	"time"
 
 	"github.com/approvers/qip/pkg/repository/dummy"
-
-	"github.com/approvers/qip/pkg/repository"
 	"github.com/approvers/qip/pkg/repository/gormRepository"
 	"github.com/approvers/qip/pkg/utils/config"
-
 	"github.com/approvers/qip/pkg/utils/token"
+	"gorm.io/driver/postgres"
+	"gorm.io/gorm"
 
+	"github.com/approvers/qip/pkg/repository"
 	"github.com/approvers/qip/pkg/server/handler/auth"
 
 	"go.uber.org/zap"
-
-	"gorm.io/driver/postgres"
-
-	"gorm.io/gorm"
-
-	"github.com/approvers/qip/pkg/server/serverErrors"
 
 	"github.com/approvers/qip/pkg/server/handler/post"
 
@@ -34,12 +28,17 @@ import (
 	"github.com/labstack/echo/v4/middleware"
 )
 
-func StartServer(port int) {
-	var userRepository repository.UserRepository
-	var postRepository repository.PostRepository
-	var fileRepository repository.FileRepository
-	var instanceRepository repository.InstanceRepository
+var (
+	userRepository     repository.UserRepository
+	postRepository     repository.PostRepository
+	fileRepository     repository.FileRepository
+	instanceRepository repository.InstanceRepository
+	userHandler        *user.Handler
+	postHandler        *post.Handler
+	authHandler        *auth.Handler
+)
 
+func initServer() {
 	if config.QipConfig.Mode != "development" {
 		db, err := gorm.Open(postgres.Open(
 			fmt.Sprintf("postgres://%s:%s@%s:%s/%s?sslmode=disable",
@@ -69,11 +68,15 @@ func StartServer(port int) {
 		userRepository = dummy.NewUserRepository(UserMockData)
 		postRepository = dummy.NewPostRepository(PostMockData)
 	}
-	userHandler := user.NewUserHandler(userRepository, fileRepository, instanceRepository)
 	key := token.SecureRandom(512)
-	postHandler := post.NewPostHandler(postRepository, key)
-	authHandler := auth.NewHandler(userRepository, key)
+	userHandler = user.NewUserHandler(userRepository, fileRepository, instanceRepository)
+	postHandler = post.NewPostHandler(postRepository, key)
+	authHandler = auth.NewHandler(userRepository, key)
 
+}
+
+func StartServer(port int) {
+	initServer()
 	e := echo.New()
 
 	logger, _ := zap.NewDevelopment()
@@ -91,16 +94,7 @@ func StartServer(port int) {
 
 	e.Use(middleware.Recover())
 	e.HTTPErrorHandler = ErrorHandler
-
-	e.POST("/api/v1/login", authHandler.LoginHandler)
-	api := e.Group("/api/v1")
-	{
-		api.Use(authHandler.TokenMiddlewareHandlerFunc)
-		api.POST("/posts", postHandler.Post)
-		api.GET("/posts/:id", postHandler.FindByID)
-
-		api.GET("/users/:id", userHandler.FindByID)
-	}
+	rootRouter(e)
 
 	go func() {
 		if err := e.Start(fmt.Sprintf(":%d", port)); err != nil && err != http.ErrServerClosed {
@@ -115,20 +109,5 @@ func StartServer(port int) {
 	defer cancel()
 	if err := e.Shutdown(ctx); err != nil {
 		e.Logger.Fatal(err)
-	}
-}
-
-func ErrorHandler(err error, c echo.Context) {
-	if h, ok := err.(*echo.HTTPError); ok {
-		if h.Code == 404 {
-			if err := c.JSON(404, serverErrors.NotFoundErrorResponseJSON); err != nil {
-				c.Logger().Error(err)
-			}
-		}
-		if h.Code == 503 {
-			if err := c.JSON(503, serverErrors.InternalErrorResponseJSON); err != nil {
-				c.Logger().Error(err)
-			}
-		}
 	}
 }
